Added some error handling code.
Added code in controllers to deal with null returned by
Model.find().exec(). Refactored some logic to accommodate these changes.
