+++ src/expressAppModules/controllers/categoriesController.js
@@ -3,7 +3,8 @@ var Category     = require('../models/categoryModel');
 
 // require mongoose and set promises to Bluebird.
 var mongoose     = require('mongoose');
-mongoose.Promise = require('bluebird');
+var Promise      = require('bluebird');
+mongoose.Promise = Promise;
 
 module.exports.get_posts = function(req, res, next) {
     var categoryId = req.params.categoryId;
@@ -14,7 +15,7 @@ module.exports.get_posts = function(req, res, next) {
 
     return Category.findById(categoryId).populate({
         path: 'posts',
-        select: '-comments -postSaveHookEnabled - postRemoveHookEnabled',
+        select: '-comments -postSaveHookEnabled -postRemoveHookEnabled',
         // I want them to be in reverse order in which thew were pushed.
         options: { limit: 20. sort: -1 },
         populate: {
@@ -28,9 +29,17 @@ module.exports.get_posts = function(req, res, next) {
             });
         }
 
+        var posts = category.posts;
+        var catPosts = [];
+        posts.forEach(function(post) {
+            if(post.user !== null) {
+                catPosts.push(post);
+            }
+        });
+
         res.json({
             entities: {
-                posts: category.posts
+                posts: catPosts
             }
         });
     }).catch(function(err) {
@@ -63,6 +72,7 @@ module.exports.get_more_posts = function(req, res, next) {
         var toSkip = cat.posts.indexOf(maxId);
         return cat.populate({
             path: 'posts',
+            select: '-comments -postSaveHookEnabled -postRemoveHookEnabled',
             options: {
                 // Watch out for the end where it starts skipping.
                 skip: toSkip,
@@ -76,7 +86,19 @@ module.exports.get_more_posts = function(req, res, next) {
                 select: '_id username miniProfilePicUrl activated -admin'
             }
         }).execPopulate().then(function(popCat) {
-            res.json(popCat.posts);
+            var catPosts = [];
+            var posts = popCat.posts;
+            posts.forEach(function(post) {
+                if(post.user !== null) {
+                    catPosts.push(post);
+                }
+            });
+
+            res.json({
+                entities: {
+                    posts: catPosts
+                }
+            });
         });
     }).catch(function(err) {
         next(err);
