+++ src/expressAppModules/controllers/relationshipsController.js
@@ -6,9 +6,6 @@ var mongoose     = require('mongoose');
 var Promise      = require('bluebird');
 mongoose.Promise = Promise;
 
-// Consider refactoring the logic of establishing relationships.
-// Might as well do it without the need to create a collection for them.
-
 // Creates and responds with new relationship.
 module.exports.create_relationship = function(req, res, next) {
     if(req.user) {
@@ -22,11 +19,14 @@ module.exports.create_relationship = function(req, res, next) {
         }
 
         var promises = [
-            User.findById(authUserId).exec(),
-            User.findById(followedId).exec()
+            User.find({ '_id': { $in: [authUserId, followedId] } }).exec()
         ];
 
         function pushRel(doc) {
+            if(owner === null) {
+                throw new Error('Cannot operate on an undefined post/user.');
+            }
+
             if(doc._id.toString() === authUserId) {
                 doc.following.push(followedId);
                 return doc.save();
@@ -44,21 +44,6 @@ module.exports.create_relationship = function(req, res, next) {
         }).catch(function(err) {
             next(err);
         });
-
-        /*
-        var relationship = new Relationship();
-        relationship.followerId = authUserId;
-        relationship.followedId = followedId;
-
-        return relationship.save().then(function(createdRel) {
-            res.json(createdRel);
-            req.relationship = createdRel;
-            next();
-        })
-        .catch(function(err) {
-            next(err);
-        });
-        */
     } else {
         // If no authenticated user
         res.status(401).json({
@@ -66,43 +51,6 @@ module.exports.create_relationship = function(req, res, next) {
         });
     }
 }
-
-/*
-// Will be substituted by post('save') hook.
-// Pushes and saves newly created relationship into
-// follower.following and followed.followers.
-module.exports.push_and_save_rel = function(req, res, next) {
-    var rel = req.relationship;
-    var followerId = rel.followerId; // ObjectId
-    var followedId = rel.followedId; // ObjectId
-
-    var promises = [
-        // What if any of these returns null? Will the Promise reject?
-        User.findById(followerId).exec(),
-        User.findById(followedId).exec()
-    ];
-
-    var promisedDocs = Promise.all(promises);
-
-    function pushIntoFollowingAndFollowers(doc) {
-        if(doc._id === followerId) {
-            doc.following.push(followedId);
-            return doc.save();
-        } else {
-            doc.followers.push(followerId);
-            return doc.save();
-        }
-    }
-
-    return promisedDocs.then(function(docs) {
-        docs.map(pushIntoFollowingAndFollowers);
-    })
-    .catch(function(err) {
-        err.logToConsole = true;
-        next(err);
-    });
-}
-*/
 
 // Deletes an existing relationship.
 module.exports.delete_relationship_follower = function(req, res, next) {
@@ -138,33 +86,6 @@ module.exports.delete_relationship_follower = function(req, res, next) {
         }).catch(function(err) {
             next(err);
         });
-
-         /*
-        return Relationship.findById(relationshipId).exec().then(function(rel) {
-            if(rel === null) {
-                res.status(404).json({
-                    message: 'Relationship not found.'
-                });
-            }
-
-            if(rel.followerId.toString() === authUserId) {
-                return rel.remove().then(function() {
-                    res.status(200).json({
-                        message: 'Relationship successfully deleted.',
-                        relationshipId: relationshipId
-                    });
-                });
-            } else {
-                // If authenticated user does not match owner of comment doc.
-                res.status(403).json({
-                    message: 'You are not authorized to perform this operation.'
-                });
-            }
-        })
-        .catch(function(err) {
-            next(err);
-        });
-        */
     } else {
         // If no authenticated user
         res.status(401).json({
